Fix async based Unit Packs buy request bug
Fixes the mistake where the user could request to buy the same unit pack multiple times but only one was allowed causing the game to go out of sync and never give the player their units.
Now the request stores a state on the server which is created when the request is made, the API call then sends that result back once, the other request will be ignored now and nothing will be sent if the API is called again by mistake.